requirements.txt
dash==2.2.0
dash-bootstrap-components==1.0.3
requests==2.27.1
python-decouple==3.6
beautifulsoup4